schwarz-precond: split LSMR into bidiag/recurrence submodules
`solve/lsmr.rs` had grown to ~1500 lines mixing the bidiagonalization
machinery, the Givens/recurrence/convergence state machines, and a
~500-line inline test module. Split along the producer/consumer seam:

- `lsmr.rs` (~90 lines) — facade with `mlsmr`, `LsmrResult`, the
  `lsmr_from_bidiag` driver, and the shared parallel-kernel constants.
- `lsmr/bidiag.rs` — the `(α, β)` stream: `Bidiagonalization` trait,
  `GolubKahan`/`ModifiedGolubKahan`, `LocalReorth`/`ModifiedLocalReorth`,
  and the vector kernels they consume.
- `lsmr/recurrence.rs` — the consumer: `Givens`, `RotationStep`,
  `LsmrRecurrenceState`, `SolutionState`, `ConvergenceState`, `Stop`.
- `lsmr/tests.rs` — the existing test suite, unchanged.

Public API frozen: `schwarz_precond::solve::lsmr::{mlsmr, LsmrResult}`
still resolves identically for `within::solver` and the
`lsmr_local_reorth` example.
